Use DocValues to calculate token size of grouped documents.

+++ core/src/main/java/nl/inl/blacklab/search/results/DocGroup.java
@@ -26,15 +26,15 @@ import nl.inl.blacklab.resultproperty.PropertyValue;
  */
 public class DocGroup extends Group<DocResult> {
     
-    public static DocGroup fromList(QueryInfo queryInfo, PropertyValue groupIdentity, List<DocResult> storedResults, int totalDocuments, int totalTokens) {
+    public static DocGroup fromList(QueryInfo queryInfo, PropertyValue groupIdentity, List<DocResult> storedResults, int totalDocuments, long totalTokens) {
         return new DocGroup(queryInfo, groupIdentity, storedResults, totalDocuments, totalTokens);
     }
     
-    private int totalTokens;
+    private long totalTokens;
 
     private int storedHits;
     
-    protected DocGroup(QueryInfo queryInfo, PropertyValue groupIdentity, List<DocResult> storedResults, int totalDocuments, int totalTokens) {
+    protected DocGroup(QueryInfo queryInfo, PropertyValue groupIdentity, List<DocResult> storedResults, int totalDocuments, long totalTokens) {
         super(groupIdentity, DocResults.fromList(queryInfo, storedResults, (SampleParameters)null, (WindowStats)null), totalDocuments);
         this.totalTokens = totalTokens;
         storedHits = 0;
@@ -52,7 +52,7 @@ public class DocGroup extends Group<DocResult> {
         return storedHits;
     }
 
-    public int totalTokens() {
+    public long totalTokens() {
         return totalTokens;
     }
